[ C7 ] - CreatingPaymentComponent
This commit marks the implementation of the PaymentButton component and Product Row which makes possible the appearance of the product in a row , a crucial addition to the application's interface. The PaymentButton component is designed to streamline the checkout process and enable users to initiate payments seamlessly. By introducing this component, the application enhances its functionality, allowing users to complete transactions efficiently and securely